+++ src/components/ui/index.ts
@@ -5,3 +5,4 @@ export * from './CornerInfoCard';
 export * from './TrendBadge';
 export * from './KpiStrip';
 export * from './DevCornerSlideHero/DevCornerSlideHero';
+export * from './LimboTicketsTable';